Tests codestyle enhancements
Models moved to separate files.
setUp method moved to base class
